refactor: features, nightly features, doc cfg
