+++ Cargo.toml
@@ -1,26 +1,23 @@
 [package]
 name = "ruint2"
-description = "Unsigned integer type with cont-generic bit length"
+description = "Unsigned integer type with const-generic bit length"
 authors = ["Remco Bloemen <remco@wicked.ventures>"]
+license = "MIT"
+rust-version = "1.65"
 version = "1.9.0"
 edition = "2021"
-homepage = "https://github.com/recmo/uint"
-repository = "https://github.com/recmo/uint"
+homepage = "https://github.com/alloy-rs/uint"
+repository = "https://github.com/alloy-rs/uint"
 keywords = ["uint"]
 categories = ["mathematics"]
-readme = "Readme.md"
-license = "MIT"
 
 [workspace]
-members = ["ruint-macro"]
+members = [".", "ruint-macro"]
+resolver = "2"
 
-[features]
-default = []
-bench = ["dep:criterion", "proptest", "rand"]
-dyn = ["dep:smallvec"]
-generic_const_exprs = []
-postgres = ["dep:postgres-types", "dep:bytes"]
-sqlx = ["dep:sqlx-core"]
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs", "--html-in-header", ".cargo/katex-header.html"]
 
 [[bench]]
 name = "criterion"
@@ -29,39 +26,40 @@ path = "criterion.rs"
 required-features = ["bench"]
 
 [dependencies]
-arbitrary = { version = "1.1", optional = true }
-ark-ff = { version = "0.4.0", optional = true }
-bn-rs = { version = "0.2", optional = true }
-bytes = { version = "1.1", optional = true }
-criterion = { version = "0.4", optional = true } # Secretly dev-dependencies
-derive_more = "0.99.17"
-fastrlp = { version = "0.3.1", optional = true }
-alloy-rlp = { version = "0.1.0", optional = true }
-num-bigint = { version = "0.4.3", optional = true }
-parity-scale-codec = { version = "3.2.1", optional = true, features = [ "max-encoded-len" ] }
-postgres-types = { version = "0.2.3", optional = true }
-primitive-types = { version = "<1", optional = true }
-proptest = { version = "1.0", optional = true }
-pyo3 = { version = "0.18.3", optional = true }
-quickcheck = { version = "1.0", optional = true }
-rand = { version = "0.8.5", optional = true }
-rlp = { version = "0.5.1", optional = true }
 ruint2-macro = { version = "1.0.3", path = "./ruint-macro" }
-serde = { version = "1.0", optional = true }
-smallvec = { version = "1.8.0", optional = true, features = [ "union" ] } # for UintDyn
-# HACK: (BLOCKED) sqlx requires a runtime to be specified.
-# <https://github.com/launchbadge/sqlx/issues/1627>
-sqlx-core = { version = "0.6", optional = true, features = [ "runtime-tokio-native-tls" ] }
+
+derive_more = "0.99.17"
 thiserror = "1.0"
-valuable = { version = "0.1.0", optional = true }
+
+# support
+alloy-rlp = { version = "0", optional = true }
+arbitrary = { version = "1", optional = true }
+ark-ff = { version = "0.4", optional = true }
+bn-rs = { version = "0.2", optional = true }
+fastrlp = { version = "0.3", optional = true }
+num-bigint = { version = "0.4", optional = true }
+parity-scale-codec = { version = "3", optional = true, features = ["derive", "max-encoded-len"] }
+primitive-types = { version = "0.12", optional = true }
+proptest = { version = "1", optional = true }
+pyo3 = { version = "0.19", optional = true }
+quickcheck = { version = "1", optional = true }
+rand = { version = "0.8", optional = true }
+rlp = { version = "0.5", optional = true }
+serde = { version = "1", optional = true }
+valuable = { version = "0.1", optional = true }
 zeroize = { version = "1.5", optional = true }
 
-[build-dependencies]
-# Rust currently has no way to conditionally compile based on compiler features.
-# `autocfg` is a workaround, but doesn't support probing for `#[features(...)]`
-# yet (there's a PR though, but crates.io doesn't allow us to use it).
-# So instead, we do the dumb thing and just check for nightly.
-rustc_version = "0.4.0"
+# postgres
+bytes = { version = "1.1", optional = true }
+postgres-types = { version = "0.2.3", optional = true }
+
+# sqlx
+# TODO: remove the feature once 0.7 is released: https://github.com/launchbadge/sqlx/issues/1627
+sqlx-core = { version = "0.6", optional = true, features = ["runtime-tokio-native-tls"] }
+
+# Secretly dev dependencies
+# TODO: Move benchmarks to benches/
+criterion = { version = "0.5", optional = true }
 
 [dev-dependencies]
 ruint2 = { path = ".", features = ["arbitrary", "proptest"] }
@@ -73,19 +71,34 @@ coverage-helper = "0.1"
 hex = "0.4.3"
 hex-literal = "0.4.1"
 postgres = "0.19.3"
-proptest = { version = "1.0" }
+proptest = "1.0"
 serde_json = "1.0"
 
-# Compilation profile for any non-workspace member.
-# Dependencies are optimized, even in a dev build. This improves dev performance
-# while having negligible impact on incremental build times.
-[profile.dev.package."*"]
-opt-level = 3
+[features]
+# nightly-only features
+nightly = []
+generic_const_exprs = ["nightly"]
 
-# Produce debug info for release builds
-[profile.release]
-debug = true
+# support
+alloy-rlp = ["dep:alloy-rlp"]
+arbitrary = ["dep:arbitrary"]
+ark-ff = ["dep:ark-ff"]
+bn-rs = ["dep:bn-rs"]
+fastrlp = ["dep:fastrlp"]
+num-bigint = ["dep:num-bigint"]
+parity-scale-codec = ["dep:parity-scale-codec"]
+primitive-types = ["dep:primitive-types"]
+proptest = ["dep:proptest"]
+pyo3 = ["dep:pyo3"]
+quickcheck = ["dep:quickcheck"]
+rand = ["dep:rand"]
+rlp = ["dep:rlp"]
+serde = ["dep:serde"]
+valuable = ["dep:valuable"]
+zeroize = ["dep:zeroize"]
 
-[package.metadata.docs.rs]
-rustdoc-args = ["--cfg", "has_cfg_doc", "--html-in-header", ".cargo/katex-header.html"]
-all-features = true
+postgres = ["dep:postgres-types", "dep:bytes"]
+sqlx = ["dep:sqlx-core"]
+
+# TODO: Move benchmarks to benches/
+bench = ["dep:criterion", "proptest", "rand"]
