+++ src/support/fastrlp.rs
@@ -1,6 +1,6 @@
 //! Support for the [`fastrlp`](https://crates.io/crates/fastrlp) crate.
 #![cfg(feature = "fastrlp")]
-#![cfg_attr(has_doc_cfg, doc(cfg(feature = "fastrlp")))]
+#![cfg_attr(docsrs, doc(cfg(feature = "fastrlp")))]
 
 use crate::Uint;
 use core::mem::size_of;
