+++ src/support/proptest.rs
@@ -1,6 +1,6 @@
 //! Support for the [`proptest`](https://crates.io/crates/proptest) crate.
 #![cfg(feature = "proptest")]
-#![cfg_attr(has_doc_cfg, doc(cfg(feature = "proptest")))]
+#![cfg_attr(docsrs, doc(cfg(feature = "proptest")))]
 
 use crate::{nlimbs, Bits, Uint};
 use proptest::{
