+++ src/support/rlp.rs
@@ -1,6 +1,6 @@
 //! Support for the [`rlp`](https://crates.io/crates/rlp) crate.
 #![cfg(feature = "rlp")]
-#![cfg_attr(has_doc_cfg, doc(cfg(feature = "rlp")))]
+#![cfg_attr(docsrs, doc(cfg(feature = "rlp")))]
 
 use crate::{Bits, Uint};
 use rlp::{Decodable, DecoderError, Encodable, Rlp, RlpStream};
