+++ src/support/zeroize.rs
@@ -2,7 +2,7 @@
 //!
 //! Currently only encodes to/from a big-endian byte array.
 #![cfg(feature = "zeroize")]
-#![cfg_attr(has_doc_cfg, doc(cfg(feature = "zeroize")))]
+#![cfg_attr(docsrs, doc(cfg(feature = "zeroize")))]
 
 use crate::{Bits, Uint};
 use zeroize::Zeroize;
